dealed with test_transforms_tensor
