+++ test/test_transforms_tensor.py
@@ -9,14 +9,15 @@ import numpy as np
 import unittest
 from typing import Sequence
 
-from common_utils import TransformsTester, get_tmp_dir, int_dtypes, float_dtypes
+from common_utils import (get_tmp_dir, int_dtypes, float_dtypes, _create_data, _create_data_batch,
+                          _assert_equal_tensor_to_pil, _assert_approx_equal_tensor_to_pil)
 from _assert_utils import assert_equal
 
 
 NEAREST, BILINEAR, BICUBIC = InterpolationMode.NEAREST, InterpolationMode.BILINEAR, InterpolationMode.BICUBIC
 
 
-class Tester(TransformsTester):
+class Tester(unittest.TestCase):
 
     def setUp(self):
         self.device = "cpu"
@@ -26,13 +27,13 @@ class Tester(TransformsTester):
             fn_kwargs = {}
 
         f = getattr(F, func)
-        tensor, pil_img = self._create_data(height=10, width=10, device=self.device)
+        tensor, pil_img = _create_data(height=10, width=10, device=self.device)
         transformed_tensor = f(tensor, **fn_kwargs)
         transformed_pil_img = f(pil_img, **fn_kwargs)
         if test_exact_match:
-            self.compareTensorToPIL(transformed_tensor, transformed_pil_img, **match_kwargs)
+            _assert_equal_tensor_to_pil(transformed_tensor, transformed_pil_img, **match_kwargs)
         else:
-            self.approxEqualTensorToPIL(transformed_tensor, transformed_pil_img, **match_kwargs)
+            _assert_approx_equal_tensor_to_pil(transformed_tensor, transformed_pil_img, **match_kwargs)
 
     def _test_transform_vs_scripted(self, transform, s_transform, tensor, msg=None):
         torch.manual_seed(12)
@@ -63,22 +64,22 @@ class Tester(TransformsTester):
         f = getattr(T, method)(**meth_kwargs)
         scripted_fn = torch.jit.script(f)
 
-        tensor, pil_img = self._create_data(26, 34, device=self.device)
+        tensor, pil_img = _create_data(26, 34, device=self.device)
         # set seed to reproduce the same transformation for tensor and PIL image
         torch.manual_seed(12)
         transformed_tensor = f(tensor)
         torch.manual_seed(12)
         transformed_pil_img = f(pil_img)
         if test_exact_match:
-            self.compareTensorToPIL(transformed_tensor, transformed_pil_img, **match_kwargs)
+            _assert_equal_tensor_to_pil(transformed_tensor, transformed_pil_img, **match_kwargs)
         else:
-            self.approxEqualTensorToPIL(transformed_tensor.float(), transformed_pil_img, **match_kwargs)
+            _assert_approx_equal_tensor_to_pil(transformed_tensor.float(), transformed_pil_img, **match_kwargs)
 
         torch.manual_seed(12)
         transformed_tensor_script = scripted_fn(tensor)
         assert_equal(transformed_tensor, transformed_tensor_script)
 
-        batch_tensors = self._create_data_batch(height=23, width=34, channels=3, num_samples=4, device=self.device)
+        batch_tensors = _create_data_batch(height=23, width=34, channels=3, num_samples=4, device=self.device)
         self._test_transform_vs_scripted_on_batch(f, scripted_fn, batch_tensors)
 
         with get_tmp_dir() as tmp_dir:
@@ -259,13 +260,13 @@ class Tester(TransformsTester):
         fn = getattr(F, func)
         scripted_fn = torch.jit.script(fn)
 
-        tensor, pil_img = self._create_data(height=20, width=20, device=self.device)
+        tensor, pil_img = _create_data(height=20, width=20, device=self.device)
         transformed_t_list = fn(tensor, **fn_kwargs)
         transformed_p_list = fn(pil_img, **fn_kwargs)
         self.assertEqual(len(transformed_t_list), len(transformed_p_list))
         self.assertEqual(len(transformed_t_list), out_length)
         for transformed_tensor, transformed_pil_img in zip(transformed_t_list, transformed_p_list):
-            self.compareTensorToPIL(transformed_tensor, transformed_pil_img)
+            _assert_equal_tensor_to_pil(transformed_tensor, transformed_pil_img)
 
         transformed_t_list_script = scripted_fn(tensor.detach().clone(), **fn_kwargs)
         self.assertEqual(len(transformed_t_list), len(transformed_t_list_script))
@@ -284,7 +285,7 @@ class Tester(TransformsTester):
         self.assertEqual(len(output), len(transformed_t_list_script))
 
         # test on batch of tensors
-        batch_tensors = self._create_data_batch(height=23, width=34, channels=3, num_samples=4, device=self.device)
+        batch_tensors = _create_data_batch(height=23, width=34, channels=3, num_samples=4, device=self.device)
         torch.manual_seed(12)
         transformed_batch_list = fn(batch_tensors)
 
@@ -350,7 +351,7 @@ class Tester(TransformsTester):
         self.assertEqual(y.shape[1], 38)
         self.assertEqual(y.shape[2], int(38 * 46 / 32))
 
-        tensor, _ = self._create_data(height=34, width=36, device=self.device)
+        tensor, _ = _create_data(height=34, width=36, device=self.device)
         batch_tensors = torch.randint(0, 256, size=(4, 3, 44, 56), dtype=torch.uint8, device=self.device)
 
         for dt in [None, torch.float32, torch.float64]:
@@ -487,7 +488,7 @@ class Tester(TransformsTester):
 
     def test_normalize(self):
         fn = T.Normalize((0.5, 0.5, 0.5), (0.5, 0.5, 0.5))
-        tensor, _ = self._create_data(26, 34, device=self.device)
+        tensor, _ = _create_data(26, 34, device=self.device)
 
         with self.assertRaisesRegex(TypeError, r"Input tensor should be a float tensor"):
             fn(tensor)
@@ -506,7 +507,7 @@ class Tester(TransformsTester):
     def test_linear_transformation(self):
         c, h, w = 3, 24, 32
 
-        tensor, _ = self._create_data(h, w, channels=c, device=self.device)
+        tensor, _ = _create_data(h, w, channels=c, device=self.device)
 
         matrix = torch.rand(c * h * w, c * h * w, device=self.device)
         mean_vector = torch.rand(c * h * w, device=self.device)
@@ -529,7 +530,7 @@ class Tester(TransformsTester):
             scripted_fn.save(os.path.join(tmp_dir, "t_norm.pt"))
 
     def test_compose(self):
-        tensor, _ = self._create_data(26, 34, device=self.device)
+        tensor, _ = _create_data(26, 34, device=self.device)
         tensor = tensor.to(dtype=torch.float32) / 255.0
 
         transforms = T.Compose([
@@ -552,7 +553,7 @@ class Tester(TransformsTester):
             torch.jit.script(t)
 
     def test_random_apply(self):
-        tensor, _ = self._create_data(26, 34, device=self.device)
+        tensor, _ = _create_data(26, 34, device=self.device)
         tensor = tensor.to(dtype=torch.float32) / 255.0
 
         transforms = T.RandomApply([
@@ -620,7 +621,7 @@ class Tester(TransformsTester):
         with self.assertRaises(ValueError, msg="If value is a sequence, it should have either a single value or 3"):
             random_erasing(img)
 
-        tensor, _ = self._create_data(24, 32, channels=3, device=self.device)
+        tensor, _ = _create_data(24, 32, channels=3, device=self.device)
         batch_tensors = torch.rand(4, 3, 44, 56, device=self.device)
 
         test_configs = [
@@ -640,7 +641,7 @@ class Tester(TransformsTester):
             scripted_fn.save(os.path.join(tmp_dir, "t_random_erasing.pt"))
 
     def test_convert_image_dtype(self):
-        tensor, _ = self._create_data(26, 34, device=self.device)
+        tensor, _ = _create_data(26, 34, device=self.device)
         batch_tensors = torch.rand(4, 3, 44, 56, device=self.device)
 
         for in_dtype in int_dtypes() + float_dtypes():
